Prevent issue with PHP warnings, when WordsMatter tried to delete unexisting output files.

+++ api/deletePost.php
@@ -9,8 +9,12 @@ if(isset($_POST['EditorId']) && isset($_POST['filename'])) {
 	if($_POST['EditorId'] == EDITOR_ID) {
 
 		// Proceed (delete the selected post).
+		// Remove input file.
 		unlink(INPUT_FOLDER . $_POST['filename'] . INPUT_EXTENSION);
-		unlink(OUTPUT_FOLDER . $_POST['filename'] . OUTPUT_EXTENSION);
+		// Remove output file (if existing).
+		if(file_exists(OUTPUT_FOLDER . $_POST['filename'] . OUTPUT_EXTENSION)) {
+			unlink(OUTPUT_FOLDER . $_POST['filename'] . OUTPUT_EXTENSION);
+		}
 
 		// Success ?
 		if(!file_exists(INPUT_FOLDER . $_POST['filename'] . INPUT_EXTENSION)
